+++ backend/Cargo.toml
@@ -17,7 +17,9 @@ awc = "2"
 env_logger = "0.8"
 futures = "0.3.1"
 bytes = "1.0.1"
-serde = { version = "1.0", features = ["derive"] }
+serde =   "1"
+serde_derive = "1"
 serde_json = "1.0"
 json = "0.12"
 listenfd = "0.3.3"
+notify = "4.0.16"
