+++ backend/src/main.rs
@@ -1,111 +1,22 @@
-use std::time::{Duration, Instant};
-
-use actix::prelude::*;
 use actix_cors::Cors;
-use actix_web::{get, http, web, App, HttpRequest, HttpResponse, HttpServer};
+use actix_web::{http, web, App, HttpRequest, HttpResponse, HttpServer};
 use actix_web::{middleware, Error};
-use actix_web_actors::ws;
 use listenfd::ListenFd;
-use serde::{Deserialize, Serialize};
 
-/// How often heartbeat pings are sent
-const HEARTBEAT_INTERVAL: Duration = Duration::from_secs(5);
-/// How long before lack of client response causes a timeout
-const CLIENT_TIMEOUT: Duration = Duration::from_secs(10);
+mod data;
+mod server;
 
 /// do websocket handshake and start `MyWebSocket` actor
 async fn ws_index(r: HttpRequest, stream: web::Payload) -> Result<HttpResponse, Error> {
-    println!("{:?}", r);
-    let res = ws::start(MyWebSocket::new(), &r, stream);
-    println!("{:?}", res);
-    res
+    server::ws_index(r, stream)
 }
 
-/// websocket connection is long running connection, it easier
-/// to handle with an actor
-struct MyWebSocket {
-    /// Client must send ping at least once per 10 seconds (CLIENT_TIMEOUT),
-    /// otherwise we drop connection.
-    hb: Instant,
-}
-
-impl Actor for MyWebSocket {
-    type Context = ws::WebsocketContext<Self>;
-
-    /// Method is called on actor start. We start the heartbeat process here.
-    fn started(&mut self, ctx: &mut Self::Context) {
-        self.hb(ctx);
-    }
-}
-
-/// Handler for `ws::Message`
-impl StreamHandler<Result<ws::Message, ws::ProtocolError>> for MyWebSocket {
-    fn handle(&mut self, msg: Result<ws::Message, ws::ProtocolError>, ctx: &mut Self::Context) {
-        // process websocket messages
-        println!("WS: {:?}", msg);
-        match msg {
-            Ok(ws::Message::Ping(msg)) => {
-                self.hb = Instant::now();
-                ctx.pong(&msg);
-            }
-            Ok(ws::Message::Pong(_)) => {
-                self.hb = Instant::now();
-            }
-            Ok(ws::Message::Text(text)) => ctx.text(text),
-            Ok(ws::Message::Binary(bin)) => ctx.binary(bin),
-            Ok(ws::Message::Close(reason)) => {
-                ctx.close(reason);
-                ctx.stop();
-            }
-            _ => ctx.stop(),
-        }
-    }
-}
-
-impl MyWebSocket {
-    fn new() -> Self {
-        Self { hb: Instant::now() }
-    }
-
-    /// helper method that sends ping to client every second.
-    ///
-    /// also this method checks heartbeats from client
-    fn hb(&self, ctx: &mut <Self as Actor>::Context) {
-        ctx.run_interval(HEARTBEAT_INTERVAL, |act, ctx| {
-            // check client heartbeats
-            if Instant::now().duration_since(act.hb) > CLIENT_TIMEOUT {
-                // heartbeat timed out
-                println!("Websocket Client heartbeat failed, disconnecting!");
-
-                // stop actor
-                ctx.stop();
-
-                // don't try to send a ping
-                return;
-            }
-
-            ctx.ping(b"");
-        });
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-struct MyJsonFile {
-    name: String,
-    number: i32,
-}
-
-async fn echo_json_file(item: web::Json<MyJsonFile>) -> HttpResponse {
-    HttpResponse::Ok().json(item.0)
+async fn echo_json_file(item: web::Json<data::MyJsonFile>) -> HttpResponse {
+    server::echo_json_file(item)
 }
 
 async fn get_json_file() -> HttpResponse {
-    let payload = MyJsonFile {
-        name: "asdf".into(),
-        number: 3,
-    };
-
-    HttpResponse::Ok().json(payload)
+    server::get_json_file()
 }
 
 #[actix_web::main]
